Relocate adapters and fragments to each package
