+++ app/src/main/java/com/wang/taipeizooguide/ui/fragment/AttractionInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.wang.taipeizooguide.ui
+package com.wang.taipeizooguide.ui.fragment
 
 import android.content.Intent
 import android.graphics.Rect
