+++ app/src/main/res/layout/fragment_zoo_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.ZooListFragment">
+    tools:context=".ui.fragment.ZooListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/zoo_recycler_view"
